+++ R/as_dmdScheme.R
@@ -11,10 +11,6 @@
 #'
 #' @rdname as_dmdScheme
 #' @export
-#' @importFrom dplyr select starts_with
-#' @importFrom methods is as
-#' @importFrom magrittr set_names %<>%
-#' @importFrom tools file_ext
 #'
 #' @examples
 #' as_dmdScheme(dmdScheme_raw(), keepData = TRUE)