+++ R/as_xml.dmdSchemeData.R
@@ -1,5 +1,4 @@
 #' @importFrom xml2 xml_new_root xml_add_child xml_attrs write_xml
-#' @importFrom tibble is_tibble
 #'
 #' @rdname as_xml
 #' @export
@@ -15,7 +14,7 @@ as_xml.dmdSchemeData <- function(
     stop("Wrong value for 'output'. 'output' has to be one of the following values:", paste(outputValues, collapse = " "))
   }
 
-  # x is of type dmdSchemeData and therefore a tibble ------------------
+  # x is of type dmdSchemeData and therefore a data.frame ------------------
 
   xml <- xml2::xml_new_root(paste0(attr(x, "propertyName"), "List"))