+++ R/validate.character.R
@@ -1,10 +1,8 @@
 #' @export
 #'
 #' @importFrom magrittr %>% %<>%
-#' @importFrom dplyr filter select
 #' @importFrom utils browseURL glob2rx
 #' @importFrom rmarkdown render
-#' @importFrom tibble tibble
 #' @importFrom utils read.csv
 #' @importFrom magrittr extract2
 #' @importFrom digest digest