+++ SCHEME_PACKAGE/install_R_package.R
@@ -1,6 +1,0 @@
-devtools::install_github(
-  repo = "Exp-Micro-Ecol-Hub/dmdScheme",
-  ref = "master",
-  force = TRUE,
-  upgrade = "never"
-)